Switch to byte literal and prevent collision
We need to switch basename to vbasename to prevent a collision on linux.
This also changes a place where we use a special dot character, which
works for latin1, but not for utf8. This replaces it with the literal
bytes to match on so it works for either latin1 or utf8