Setup Clang as compiler
Clang is significantly more strict of a compiler than GCC - but has
added benefits that LLVM is significantly better at optimizations. This
aligns the project along a requirement that clang has. Clang requires
that if there is ever a DLL_EXPORT being applied to a type declared in
either a c file, or a header file, that DLL_EXPORT must be on *both*
declarations and not only one. Clang claims that this prevents linking
errors where a type isn't as available as the author believes it is.

Clang builds also require that strings be valid utf-8. Previously there
were usage of special characters that fit in valid latin1 strings. This
aligns them to use their literal byte codes, and adds a change to
explicitly declare that a pointer is volatile (a clang requirement).