Fix memory bug in new malloc wrapper
This fixes a sneaky memory bug in the new malloc wrapper. Previously,
when we used windows heap allocator directly, we asked for zero'd bytes.
Calling malloc on it's own does not zero the bytes returned. This lead
to a bug when testing, where mousing over items in the inventory would
not highlight them, and they were not selectable. This was because the
buffer used for buttons "but" was filled with junk bytes that were being
interpreted as real data, but with values that made no sense for
determining the button hitbox for items in the inventory.

It also caused the quest log to miraculously appear when clicking random
points around the inventory.

Calloc will correctly zero the bytes for us, ensuring the behavior stays
the same, and upon testing this fixed the bug.