Fix 32 bit pointer math bug on 64 bit
This logic worked fine in 32 bit builds, but creates a bug on 64 bit
builds. In 64 bit dllist and rem are 64 bit addresses, not 32 bit
addresses, so (unsigned char*)dllist - (unsigned char*)rem is a 64 bit
ptrdiff_t valule that we are stuffin into a 32 bit int. If that diff is
larger than what would fit in 32 bits it truncates, which leads to the
wrong offset being added to a valid pointer - which leads to garbage
pointers in dlsort.

This was caught by launching the game with an ASan build.