Fix left shift into sign bit
When running a build of the game with zig as the C compiler, with the
sanitize_c setting set to full, we get a detection on launch that when
IRGBA get's used, then a left shift of the alpha channel by 24 shifts it
into the sign bit, which is undefined behavior. This casts each channel
to unsigned before shifting so a left shift by 24 is always
representable, assuming values are always constrained to 0-255.